fix(#56): accumulate DotProduct in double, matching CosineSimilarity
DotProduct now uses double accumulation with a single (float) cast on
return, consistent with CosineSimilarity. Reduces rounding error over
high-dimension (1536+) embeddings when scoring against pre-normalized
vectors.

+++ src/Cisharpai.Rag/VectorMath.cs
@@ -14,10 +14,10 @@ public static class VectorMath
     {
         ThrowIfDimensionMismatch(a.Length, b.Length);
 
-        var sum = 0f;
+        var sum = 0.0;
         for (var i = 0; i < a.Length; i++)
-            sum += a[i] * b[i];
-        return sum;
+            sum += (double)a[i] * b[i];
+        return (float)sum;
     }
 
     /// <inheritdoc cref="DotProduct(ReadOnlySpan{float}, ReadOnlySpan{float})"/>
